# docs/swagger.yaml
basePath: /
definitions:
  internal_auth.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_auth.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/internal_auth.User'
    type: object
  internal_auth.RegisterRequest:
    properties:
      company_name:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  internal_auth.SystemAnalytics:
    properties:
      active_users:
        type: integer
      inactive_users:
        type: integer
      total_users:
        type: integer
    type: object
  internal_auth.User:
    properties:
      atualizado_em:
        type: string
      criado_em:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        description: ADMIN ou USER
        type: string
      tenant_id:
        type: string
    type: object
  internal_auth.toggleRequest:
    properties:
      active:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Suporte JR DEV
    url: https://github.com/jrdev-3
  description: API de backend para o sistema de Mini ERP, desenvolvida em Go seguindo
    Clean Architecture.
  termsOfService: http://swagger.io/terms/
  title: Mini ERP API
  version: "1.0"
paths:
  /api/v1/admin/users:
    get:
      consumes:
      - application/json
      description: Retorna a lista completa de todos os colaboradores pertencentes
        ao mesmo tenant do administrador (Requer ADMIN).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_auth.User'
            type: array
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso proibido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Listar funcionários do comércio (Administrador do Comércio)
      tags:
      - administracao
    post:
      consumes:
      - application/json
      description: Cadastra um novo colaborador com papel de USER vinculado ao mesmo
        tenant do administrador criador (Requer ADMIN).
      parameters:
      - description: Dados do funcionário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.User'
        "400":
          description: Dados inválidos ou e-mail já cadastrado
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso proibido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Cadastrar funcionário (Administrador do Comércio)
      tags:
      - administracao
  /api/v1/admin/users/{id}/toggle:
    patch:
      consumes:
      - application/json
      description: Altera o status de atividade de um usuário dentro de seu próprio
        tenant (Requer ADMIN).
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: string
      - description: Status ativo (active)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.toggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Status atualizado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso proibido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Ativar ou desativar um usuário
      tags:
      - administracao
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Valida as credenciais do usuário e retorna um token JWT de acesso
        offline.
      parameters:
      - description: Credenciais de login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.LoginResponse'
        "400":
          description: E-mail ou senha obrigatórios
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Credenciais inválidas
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Usuário inativo
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Realizar autenticação de usuário
      tags:
      - autenticacao
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Cria uma nova conta de inquilino (empresa) e o primeiro usuário
        com cargo de ADMIN.
      parameters:
      - description: Dados de cadastro
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.User'
        "400":
          description: Dados inválidos ou e-mail já cadastrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar novo Tenant e Administrador
      tags:
      - autenticacao
  /api/v1/system/analytics:
    get:
      consumes:
      - application/json
      description: Retorna o total de usuários ativos, inativos e empresas cadastradas
        no ecossistema (Requer SUPER_ADMIN).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.SystemAnalytics'
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso proibido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obter métricas de saúde da plataforma (Super Admin)
      tags:
      - administracao-plataforma
  /api/v1/system/users:
    get:
      consumes:
      - application/json
      description: Retorna a lista completa de todos os usuários cadastrados no ecossistema
        (Requer SUPER_ADMIN).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_auth.User'
            type: array
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso proibido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Listar todos os usuários da plataforma (Super Admin)
      tags:
      - administracao-plataforma
  /api/v1/system/users/{id}/toggle:
    patch:
      consumes:
      - application/json
      description: Altera o status de atividade de qualquer usuário de qualquer tenant
        na plataforma (Requer SUPER_ADMIN).
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: string
      - description: Status ativo (active)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.toggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Status do usuário atualizado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Acesso proibido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Ativar ou desativar qualquer usuário (Super Admin)
      tags:
      - administracao-plataforma
securityDefinitions:
  ApiKeyAuth:
    description: 'Insira o token JWT no formato: Bearer <TOKEN>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"